+++ src/Products/MailHost/MailHost.py
@@ -417,16 +417,31 @@ def _mungeHeaders(messageText, mto=None, mfrom=None, subject=None,
         # we don't use get_content_type because that has a default
         # value of 'text/plain'
         mo.set_type(msg_type)
-    charset_match = CHARSET_RE.search(mo['Content-Type'] or '')
-    if charset and not charset_match:
-        # Don't change the charset if already set
-        # This encodes the payload automatically based on the default
-        # encoding for the charset
-        mo.set_charset(charset)
-    elif charset_match and not charset:
-        # If a charset parameter was provided use it for header encoding below,
-        # Otherwise, try to use the charset provided in the message.
-        charset = charset_match.groups()[0]
+    if not mo.is_multipart():
+        charset_match = CHARSET_RE.search(mo['Content-Type'] or '')
+        if charset and not charset_match:
+            # Don't change the charset if already set
+            # This encodes the payload automatically based on the default
+            # encoding for the charset
+            mo.set_charset(charset)
+        elif charset_match and not charset:
+            # If a charset parameter was provided use it for header encoding below,
+            # Otherwise, try to use the charset provided in the message.
+            charset = charset_match.groups()[0]
+    else:
+        # Do basically the same for each payload as for the complete
+        # multipart message.
+        for index, payload in enumerate(mo.get_payload()):
+            if not isinstance(payload, Message):
+                payload = message_from_string(payload)
+            charset_match = CHARSET_RE.search(payload['Content-Type'] or '')
+            if payload.get_filename() is None:
+                # No binary file
+                if charset and not charset_match:
+                    payload.set_charset(charset)
+                elif charset_match and not charset:
+                    charset = charset_match.groups()[0]
+            mo.get_payload()[index] = payload
 
     # Parameters given will *always* override headers in the messageText.
     # This is so that you can't override or add to subscribers by adding